# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  email: string;
  name: string;
  picture?: string;
}

export interface FreeUsage {
  authenticated: boolean;
  usedOperations: number;
  limitOperations: number;
  remainingOperations: number;
  unlimited?: boolean;
}

export interface HealthStatus {
  status: string;
}

export interface Activity {
  id: string;
  label: string;
  detail: string;
  timestamp: string;
  tone: string;
}

export interface WorkspaceSummary {
  documents: number;
  activeJobs: number;
  wordsProcessed: number;
  intelligenceScore: number;
  credits: number;
  recentActivity: Activity[];
}

export interface Document {
  id: string;
  title: string;
  excerpt: string;
  words: number;
  score: number;
  updatedAt: string;
  status: string;
}

export interface DocumentInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
}

export interface GenerationJob {
  id: string;
  title: string;
  feature: string;
  provider: string;
  progress: number;
  status: string;
  createdAt: string;
}

export interface GenerationJobInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  feature: string;
  /** @minLength 1 */
  provider: string;
}

export interface DiagnosticInput {
  /** @minLength 1 */
  functionId: string;
  /** @minLength 1 */
  provider: string;
  /** @minLength 1 */
  input: string;
  thinker?: string;
  /**
     * @minimum 0
     * @maximum 20
     */
  quotes?: number;
}

export interface DiagnosticCheck {
  id: string;
  title: string;
  status: string;
  input: string;
  output: string;
  providerUsed: string;
  elapsedMs: number;
}

export interface DiagnosticReport {
  runId: string;
  checks: DiagnosticCheck[];
}

export interface AiDetectionInput {
  /** @minLength 1 */
  text: string;
  /** @minLength 1 */
  provider: string;
}

export interface AiDetectionResult {
  /**
     * @minimum 0
     * @maximum 100
     */
  aiProbability: number;
  label: string;
  explanation: string;
  providerUsed: string;
}

export type CoherenceInputMode = typeof CoherenceInputMode[keyof typeof CoherenceInputMode];


export const CoherenceInputMode = {
  autodetect: 'autodetect',
  logical: 'logical',
  mathematical: 'mathematical',
  persuasive: 'persuasive',
  scientific: 'scientific',
  philosophical: 'philosophical',
  structural: 'structural',
  narrative: 'narrative',
  legal: 'legal',
  technical: 'technical',
} as const;

export interface CoherenceInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  text: string;
  instructions?: string;
  targetWords?: string;
  /** @minLength 1 */
  provider: string;
  mode: CoherenceInputMode;
  allowFallback?: boolean;
  thinker?: string;
  /**
     * @minimum 0
     * @maximum 20
     */
  quotes?: number;
}

export interface CoherenceDimension {
  label: string;
  score: number;
  note: string;
}

export interface CoherenceScorecard {
  mode: string;
  /** @nullable */
  score: number | null;
  verdict: string;
  dimensions: CoherenceDimension[];
}

export interface CoherenceProviderUsage {
  evidence: string[];
  skeleton: string[];
  audit: string[];
}

export interface CoherenceResult {
  jobId: string;
  status: string;
  providerUsed: string;
  skeleton: string;
  output: string;
  /** @nullable */
  coherenceScore: number | null;
  scorecard: CoherenceScorecard;
  providerUsage: CoherenceProviderUsage;
}

export interface DocumentFunctionInput {
  /** @minLength 1 */
  functionId: string;
  text: string;
  instructions?: string;
  targetWords?: string;
  /** @minLength 1 */
  provider: string;
  thinker?: string;
  /**
     * @minimum 0
     * @maximum 20
     */
  quotes?: number;
  resumeDocument?: string;
  resumeSkeleton?: string;
}

export interface DocumentFunctionResult {
  output: string;
  providerUsed: string;
}

export interface OcrInput {
  /** @minLength 1 */
  imageData: string;
}

export interface OcrResult {
  text: string;
}

export type GetCurrentUser200 = {
  user: AuthUser | null;
};