# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diagnostics/run:
    post:
      operationId: runDiagnostics
      tags: [workspace]
      summary: Run live platform diagnostics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosticInput"
      responses:
        "200":
          description: Diagnostic report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticReport"
...
  /coherence/analyze:
    post:
      operationId: analyzeCoherence
      tags: [workspace]
      summary: Analyze or reconstruct a document with cross-chunk coherence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoherenceInput"
      responses:
        "200":
          description: Coherence result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoherenceResult"
  /functions/run:
    post:
      operationId: runDocumentFunction
      tags: [workspace]
      summary: Run a document portal operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentFunctionInput"
      responses:
        "200":
          description: Function result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentFunctionResult"
...
components:
  schemas:
...
    DiagnosticInput:
      type: object
      properties:
        functionId: { type: string, minLength: 1 }
        provider: { type: string, minLength: 1 }
        input: { type: string, minLength: 1 }
        thinker: { type: string }
        quotes: { type: number, minimum: 0, maximum: 20 }
      required: [functionId, provider, input]
    DiagnosticCheck:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        status: { type: string }
        input: { type: string }
        output: { type: string }
        providerUsed: { type: string }
        elapsedMs: { type: number }
      required: [id, title, status, input, output, providerUsed, elapsedMs]
    DiagnosticReport:
      type: object
      properties:
        runId: { type: string }
        checks:
          type: array
          items:
            $ref: "#/components/schemas/DiagnosticCheck"
      required: [runId, checks]
...
    CoherenceInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        text: { type: string, minLength: 1 }
        instructions: { type: string }
        targetWords: { type: string }
        provider: { type: string, minLength: 1 }
        mode:
          type: string
          enum: [autodetect, logical, mathematical, persuasive, scientific, philosophical, structural, narrative, legal, technical]
        allowFallback:
          type: boolean
          default: true
        thinker: { type: string }
        quotes: { type: number, minimum: 0, maximum: 20 }
      required: [title, text, provider, mode]
    CoherenceDimension:
      type: object
      properties:
        label: { type: string }
        score: { type: number }
        note: { type: string }
      required: [label, score, note]
    CoherenceScorecard:
      type: object
      properties:
        mode: { type: string }
        score: { type: [number, "null"] }
        verdict: { type: string }
        dimensions:
          type: array
          items:
            $ref: "#/components/schemas/CoherenceDimension"
      required: [mode, score, verdict, dimensions]
    CoherenceProviderUsage:
      type: object
      properties:
        evidence:
          type: array
          items: { type: string }
        skeleton:
          type: array
          items: { type: string }
        audit:
          type: array
          items: { type: string }
      required: [evidence, skeleton, audit]
    CoherenceResult:
      type: object
      properties:
        jobId: { type: string }
        status: { type: string }
        providerUsed: { type: string }
        skeleton: { type: string }
        output: { type: string }
        coherenceScore: { type: [number, "null"] }
        scorecard:
          $ref: "#/components/schemas/CoherenceScorecard"
        providerUsage:
          $ref: "#/components/schemas/CoherenceProviderUsage"
      required: [jobId, status, providerUsed, skeleton, output, coherenceScore, scorecard, providerUsage]
    DocumentFunctionInput:
      type: object
      properties:
        functionId: { type: string, minLength: 1 }
        text: { type: string }
        instructions: { type: string }
        targetWords: { type: string }
        provider: { type: string, minLength: 1 }
        thinker: { type: string }
        quotes: { type: number, minimum: 0, maximum: 20 }
        resumeDocument: { type: string }
        resumeSkeleton: { type: string }
      required: [functionId, text, provider]
    DocumentFunctionResult:
      type: object
      properties:
        output: { type: string }
        providerUsed: { type: string }
      required: [output, providerUsed]
...