# lib/api-zod/src/generated/types/coherenceDimension.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CoherenceDimension {
  label: string;
  score: number;
  note: string;
}