# lib/api-zod/src/generated/types/coherenceResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoherenceProviderUsage } from './coherenceProviderUsage';
import type { CoherenceScorecard } from './coherenceScorecard';

export interface CoherenceResult {
  jobId: string;
  status: string;
  providerUsed: string;
  skeleton: string;
  output: string;
  /** @nullable */
  coherenceScore: number | null;
  scorecard: CoherenceScorecard;
  providerUsage: CoherenceProviderUsage;
}