# lib/api-zod/src/generated/types/coherenceScorecard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoherenceDimension } from './coherenceDimension';

export interface CoherenceScorecard {
  mode: string;
  /** @nullable */
  score: number | null;
  verdict: string;
  dimensions: CoherenceDimension[];
}