# lib/api-zod/src/generated/types/diagnosticCheck.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiagnosticCheck {
  id: string;
  title: string;
  status: string;
  input: string;
  output: string;
  providerUsed: string;
  elapsedMs: number;
}