# lib/api-zod/src/generated/types/diagnosticReport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnosticCheck } from './diagnosticCheck';

export interface DiagnosticReport {
  runId: string;
  checks: DiagnosticCheck[];
}