# lib/api-zod/src/generated/types/documentFunctionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentFunctionInput {
  /** @minLength 1 */
  functionId: string;
  text: string;
  instructions?: string;
  targetWords?: string;
  /** @minLength 1 */
  provider: string;
  thinker?: string;
  /**
     * @minimum 0
     * @maximum 20
     */
  quotes?: number;
  resumeDocument?: string;
  resumeSkeleton?: string;
}